+++ CHANGES.rst
@@ -6,6 +6,9 @@ There's a frood who really knows where his towel is.
 1.0a10 (unreleased)
 ^^^^^^^^^^^^^^^^^^^
 
+- A new behavior was included to add the possibility of having a background image on the default view of the item.
+  [hvelarde]
+
 - Fix modification time disappearing from context data and handling missing value (fixes `#449`_).
   [mhora]